refactor: splitted up into multiple files, added more function descriptions fix: warnigns and errors as info if not critical
